# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sarangi DMS API
  version: 1.0.0-rc2
  description: Backend API for the Sarangi Dental Management System.
  contact:
    name: Sarangi DMS Support
    email: support@example.com
servers:
  - url: https://api.dev.example.com
    description: Dev
  - url: https://api.staging.example.com
    description: Staging
  - url: https://api.example.com
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /patients/{patientId}/visits:
    get:
      tags: [Patients, Visits]
      summary: List visits for a patient
      description: Lists all visits associated with a given patient.
      operationId: listPatientVisits
      parameters:
        - $ref: '#/components/parameters/PatientId'
      responses:
        '200':
          description: Visit list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Visit'
        '400':
          description: Invalid patient id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /visits:
    post:
      tags: [Visits]
      summary: Create visit
      description: Creates a new visit in QUEUED status.
      operationId: createVisit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitCreate'
      responses:
        '201':
          description: Visit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /visits/queue:
    get:
      tags: [Visits]
      summary: Get doctor queue
      description: >
        Returns the queue for a given doctor and date, ordered FIFO by creation time.
      operationId: getDoctorQueue
      parameters:
        - in: query
          name: doctorId
          required: true
          schema:
            type: string
          description: Doctor user id.
        - in: query
          name: date
          required: false
          schema:
            type: string
            format: date
          description: Visit date (YYYY-MM-DD). Defaults to today on backend.
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [QUEUED, IN_PROGRESS, DONE]
          description: Optional visit status filter.
      responses:
        '200':
          description: Queue items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Visit'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /visits/queue/take-seat:
    post:
      tags: [Visits]
      summary: Promote visit from QUEUED to IN_PROGRESS
      description: >
        Dedicated "take seat" operation. Enforces that the doctor is not already busy.
      operationId: takeSeat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [visitId]
              properties:
                visitId:
                  type: string
      responses:
        '200':
          description: Visit updated to IN_PROGRESS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid status transition or doctor already busy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /visits/{visitId}:
    get:
      tags: [Visits]
      summary: Get visit by id
      description: Returns a single visit.
      operationId: getVisitById
      parameters:
        - $ref: '#/components/parameters/VisitId'
      responses:
        '200':
          description: Visit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '400':
          description: Invalid visit id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /visits/{visitId}/status:
    patch:
      tags: [Visits]
      summary: Update visit status
      description: >
        Updates the visit status (QUEUED → IN_PROGRESS → DONE) with state-machine checks.
      operationId: updateVisitStatus
      parameters:
        - $ref: '#/components/parameters/VisitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitStatusUpdate'
      responses:
        '200':
          description: Visit updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid status transition or doctor already busy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /visits/{visitId}/followup:
    get:
      tags: [Visits]
      summary: Get follow-up for visit
      description: Returns the follow-up configuration for a visit.
      operationId: getVisitFollowup
      parameters:
        - $ref: '#/components/parameters/VisitId'
      responses:
        '200':
          description: Follow-up
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowUp'
        '400':
          description: Invalid visit id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Follow-up not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Visits]
      summary: Upsert follow-up for visit
      description: >
        Creates or updates follow-up details for a visit.
      operationId: upsertVisitFollowup
      parameters:
        - $ref: '#/components/parameters/VisitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowUpUpsert'
      responses:
        '200':
          description: Follow-up upserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowUp'
        '400':
          description: Validation error or follow-up rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Visit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /visits/{visitId}/followup/status:
    patch:
      tags: [Visits]
      summary: Update follow-up status
      description: Updates only the follow-up status (ACTIVE/COMPLETED/CANCELLED).
      operationId: updateVisitFollowupStatus
      parameters:
        - $ref: '#/components/parameters/VisitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowUpStatusUpdate'
      responses:
        '200':
          description: Follow-up updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowUp'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Follow-up not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /visits/{visitId}/rx:
    post:
      tags: [Prescriptions, Visits]
      summary: Create prescription for visit
      description: >
        Creates a new prescription version for a visit and stores the full JSON in S3.
      operationId: createVisitPrescription
      parameters:
        - $ref: '#/components/parameters/VisitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lines]
              properties:
                lines:
                  type: array
                  minItems: 1
                  items:
                    $ref: '#/components/schemas/RxLine'
      responses:
        '201':
          description: Prescription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RxCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Visit or patient not found/deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /visits/{visitId}/xrays:
    post:
      tags: [Xrays, Visits]
      summary: Attach X-ray metadata to visit
      description: >
        Records X-ray metadata for a visit, given an already-uploaded object key.
      operationId: createVisitXrayMetadata
      parameters:
        - $ref: '#/components/parameters/VisitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/XrayMetadataInput'
      responses:
        '201':
          description: X-ray metadata created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Xray'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Visit or patient not found/deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: X-ray id conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /visits/{visitId}/checkout:
    post:
      tags: [Billing, Visits]
      summary: Checkout visit and create bill
      description: >
        Idempotent per visit. Only allowed when visit status is DONE. Optionally creates a follow-up.
      operationId: checkoutVisit
      parameters:
        - $ref: '#/components/parameters/VisitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillingCheckoutInput'
      responses:
        '201':
          description: Billing record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Billing'
        '400':
          description: Billing or follow-up rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate checkout or visit not DONE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /visits/{visitId}/bill:
    get:
      tags: [Billing, Visits]
      summary: Get billing details for visit
      description: Returns the billing record for a visit, if present.
      operationId: getVisitBilling
      parameters:
        - $ref: '#/components/parameters/VisitId'
      responses:
        '200':
          description: Billing record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Billing'
        '400':
          description: Invalid visit id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Billing record not found or visit/patient deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PatientId:
      in: path
      name: patientId
      required: true
      schema:
        type: string
    VisitId:
      in: path
      name: visitId
      required: true
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
          description: Human-readable error message.
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Optional field-level validation errors.
        traceId:
          type: string
...
    VisitCreate:
      type: object
      required: [patientId, doctorId, reason]
      properties:
        patientId:
          type: string
        doctorId:
          type: string
        reason:
          type: string
    VisitStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [QUEUED, IN_PROGRESS, DONE]
    Visit:
      type: object
      required: [visitId, patientId, doctorId, reason, status, visitDate, createdAt, updatedAt]
      properties:
        visitId:
          type: string
        patientId:
          type: string
        doctorId:
          type: string
        reason:
          type: string
        status:
          type: string
          enum: [QUEUED, IN_PROGRESS, DONE]
        visitDate:
          type: string
          format: date
        createdAt:
          type: integer
        updatedAt:
          type: integer
        billingAmount:
          type: number
    FollowUp:
      type: object
      required: [visitId, followUpDate, contactMethod, status, createdAt, updatedAt]
      properties:
        visitId:
          type: string
        followUpDate:
          type: string
          format: date
        reason:
          type: string
        contactMethod:
          type: string
          enum: [CALL, SMS, WHATSAPP, OTHER]
        status:
          type: string
          enum: [ACTIVE, COMPLETED, CANCELLED]
        createdAt:
          type: integer
        updatedAt:
          type: integer
    FollowUpUpsert:
      type: object
      required: [followUpDate]
      properties:
        followUpDate:
          type: string
          format: date
        reason:
          type: string
        contactMethod:
          type: string
          enum: [CALL, SMS, WHATSAPP, OTHER]
    FollowUpStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ACTIVE, COMPLETED, CANCELLED]
...
    BillingCheckoutInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BillingLine'
        discountAmount:
          type: number
          minimum: 0
          default: 0
        taxAmount:
          type: number
          minimum: 0
          default: 0
        followUp:
          $ref: '#/components/schemas/CheckoutFollowUpInput'
...
    XrayMetadataInput:
      type: object
      required: [xrayId, contentType, size, takenAt, takenByUserId]
      properties:
        xrayId:
          type: string
        contentType:
          type: string
          enum: [image/jpeg, image/png]
        size:
          type: integer
          minimum: 1024
          maximum: 10485760
        takenAt:
          type: integer
        takenByUserId:
          type: string
    Xray:
      type: object
      required: [xrayId, visitId, contentKey, contentType, size, takenAt, takenByUserId, createdAt]
      properties:
        xrayId:
          type: string
        visitId:
          type: string
        contentKey:
          type: string
        thumbKey:
          type: string
        contentType:
          type: string
          enum: [image/jpeg, image/png]
        size:
          type: integer
        takenAt:
          type: integer
        takenByUserId:
          type: string
        createdAt:
          type: integer
        deletedAt:
          type: integer
...
    RxLine:
      type: object
      required: [medicine, dose, frequency, duration]
      properties:
        medicine:
          type: string
        dose:
          type: string
        frequency:
          type: string
          enum: [QD, BID, TID, QID, HS, PRN]
        duration:
          type: integer
          minimum: 1
          maximum: 365
        sig:
          type: string
        timing:
          type: string
          enum: [BEFORE_MEAL, AFTER_MEAL, ANY]
        notes:
          type: string
    RxCreateResponse:
      type: object
      required: [rxId, visitId, version, createdAt, updatedAt]
      properties:
        rxId:
          type: string
        visitId:
          type: string
        version:
          type: integer
        createdAt:
          type: integer
        updatedAt:
          type: integer
...